Removed serverles defaults from the configuration.
This version also includes an additional API call for getting a bucket, via bucket() which will return a bucket wrapped in a promise for call chainging.

Changed the default behavior for loading a piece of data that doesnt exist to not reject the promise. New configuration added to change the default.